requirements.txt
https://www.djangoproject.com/download/1.5b2/tarball/
dj-database-url==0.2.1
django-model-utils==1.2.0
gunicorn==0.15.0
newrelic==1.8.0.13
psycopg2==2.4.5
raven==2.0.10
